# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BPMS API",
        "description": "API documentation for BPMS app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/image": {
            "post": {
                "tags": [
                    "Users Images"
                ],
                "summary": "Create or update image data",
                "description": "Creates a new image record or updates an existing one based on the 'bp' value.",
                "operationId": "1b30a2e33466a980719efddab96aa395",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bp",
                                    "image_data"
                                ],
                                "properties": {
                                    "bp": {
                                        "type": "string",
                                        "example": "1300001000"
                                    },
                                    "image_data": {
                                        "type": "string",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUA..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data saved successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Employee not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/image/{bp}": {
            "get": {
                "tags": [
                    "Users Images"
                ],
                "summary": "Get image data by BP",
                "description": "Returns image data for the given BP value.",
                "operationId": "f9bec7bce053377718ac16790b7348f1",
                "parameters": [
                    {
                        "name": "bp",
                        "in": "path",
                        "description": "BP value",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1300001000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "bp": {
                                                    "type": "string",
                                                    "example": "1300001000"
                                                },
                                                "image_data": {
                                                    "type": "string",
                                                    "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUA..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "BP not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "BP not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}